grai-client/pyproject.toml
[tool.poetry]
name = "grai-client"
version = "0.1.0"
description = ""
authors = ["Ian Eaves <ian.k.eaves@gmail.com>"]
license = "Elastic-2.0"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.1"
requests = "^2.28.1"
PyYAML = "^6.0"
multimethod = "^1.8"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.971"
isort = "^5.10.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
